Move yolov6l from experimental to demos
