+++ models/demos/yolov6l/demo/images/bus.jpg
