+++ models/demos/yolov6l/demo/images/elephants.jpg
