+++ models/demos/yolov6l/runner/performant_runner.py
@@ -4,7 +4,7 @@
 
 
 import ttnn
-from models.experimental.yolov6l.runner.performant_runner_infra import YOLOv6lPerformanceRunnerInfra
+from models.demos.yolov6l.runner.performant_runner_infra import YOLOv6lPerformanceRunnerInfra
 from tests.ttnn.utils_for_testing import assert_with_pcc
