+++ models/demos/yolov6l/tt/ttnn_bepc3.py
@@ -3,8 +3,8 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import ttnn
-from models.experimental.yolov6l.tt.ttnn_repblock import TtRepBlock
-from models.experimental.yolov6l.tt.common import Yolov6l_Conv2D
+from models.demos.yolov6l.tt.common import Yolov6l_Conv2D
+from models.demos.yolov6l.tt.ttnn_repblock import TtRepBlock
 
 
 class TtBepC3:
