+++ models/demos/yolov6l/weights_download.sh
