+++ tests/ttnn/integration_tests/yolov6l/test_ttnn_yolov6l.py
@@ -7,8 +7,8 @@ import pytest
 import torch
 
 import ttnn
-from models.experimental.yolov6l.tt.model_preprocessing import create_yolov6l_model_parameters, load_torch_model_yolov6l
-from models.experimental.yolov6l.tt.ttnn_yolov6l import TtYolov6l
+from models.demos.yolov6l.tt.model_preprocessing import create_yolov6l_model_parameters, load_torch_model_yolov6l
+from models.demos.yolov6l.tt.ttnn_yolov6l import TtYolov6l
 from tests.ttnn.utils_for_testing import assert_with_pcc
